+++ src/mod/DictionaryList.js
@@ -1,6 +1,6 @@
 import _func from 'complex-func'
 import DefaultDataWithLife from './../data/DefaultDataWithLife'
-import DictionaryData from './DictionaryData'
+import DictionaryItem from './DictionaryItem'
 import OptionData from './OptionData'
 import LayoutData from './LayoutData'
 
@@ -87,7 +87,7 @@ class DictionaryList extends DefaultDataWithLife {
   }
   /**
    * 加载默认初始值.子类自动按照父类来源设置
-   * @param {object} optiondata DictionaryData初始化参数
+   * @param {object} optiondata DictionaryItem初始化参数
    * @param {*} parentData 父元素
    * @param {*} isChildren 是否是子类
    */
@@ -147,7 +147,7 @@ class DictionaryList extends DefaultDataWithLife {
         if (act.build) {
           // 构建字典数据
           ditemOption.parent = this
-          ditem = new DictionaryData(ditemOption, {
+          ditem = new DictionaryItem(ditemOption, {
             layout: this.getLayout()
           })
           this.data.set(ditem.prop, ditem)
@@ -186,7 +186,7 @@ class DictionaryList extends DefaultDataWithLife {
   }
   /**
    * 解析字典初始化数据,获取子字典创建模式
-   * @param {DictionaryData} ditem 对应字典实例
+   * @param {DictionaryItem} ditem 对应字典实例
    * @param {object} originOption 字典初始化数据
    * @returns {'' | 'self' | 'build'}
    */
@@ -208,7 +208,7 @@ class DictionaryList extends DefaultDataWithLife {
   }
   /**
    * 创建字典的子字典列表
-   * @param {DictionaryData} ditem 对应字典实例
+   * @param {DictionaryItem} ditem 对应字典实例
    * @param {object} originOption 字典初始化数据
    * @param {boolean} isChildren 是否子类
    */
@@ -260,7 +260,7 @@ class DictionaryList extends DefaultDataWithLife {
 
   /**
    * 获取列表MAP
-   * @returns {Map<DictionaryData>}
+   * @returns {Map<DictionaryItem>}
    */
   getList () {
     return this.data
@@ -270,7 +270,7 @@ class DictionaryList extends DefaultDataWithLife {
    * 获取字典对象
    * @param {*} data 值
    * @param {'prop' | 'id'} from 获取类型
-   * @returns {DictionaryData}
+   * @returns {DictionaryItem}
    */
   getItem (data, from = 'prop') {
     if (from == 'prop') {
@@ -366,7 +366,7 @@ class DictionaryList extends DefaultDataWithLife {
   }
   /**
    * 格式化数据
-   * @param {DictionaryData} ditem 字典
+   * @param {DictionaryItem} ditem 字典
    * @param {object} targetitem 目标数据
    * @param {object} originitem 源数据
    * @param {string} originfromType 来源originfromType
@@ -442,17 +442,17 @@ class DictionaryList extends DefaultDataWithLife {
   /**
    * 获取符合模块要求的字典列表
    * @param {string} mod 模块名称
-   * @returns {DictionaryData[]}
+   * @returns {DictionaryItem[]}
    */
   getModList (modType) {
     return this.getModListNext([], this.data, modType)
   }
   /**
    * 从dataMap获取符合模块要求的字典列表
-   * @param {DictionaryData[]} modList 返回的字典列表
-   * @param {Map<DictionaryData>} dataMap 字典Map
+   * @param {DictionaryItem[]} modList 返回的字典列表
+   * @param {Map<DictionaryItem>} dataMap 字典Map
    * @param {string} modType 模块名称
-   * @returns {DictionaryData[]}
+   * @returns {DictionaryItem[]}
    */
   getModListNext (modList, dataMap, modType) {
     for (let ditem of dataMap.values()) {
@@ -476,7 +476,7 @@ class DictionaryList extends DefaultDataWithLife {
   /**
    * 将模块列表根据payload转换为页面需要数据的列表
    * @param {string} modType 模块名称
-   * @param {DictionaryData[]} modlist 模块列表
+   * @param {DictionaryItem[]} modlist 模块列表
    * @param {object} [payload] 参数
    * @returns {*[]}
    */
@@ -498,7 +498,7 @@ class DictionaryList extends DefaultDataWithLife {
   }
   /**
    * 根据模块列表生成对应的form对象
-   * @param {DictionaryData[]} modlist 模块列表
+   * @param {DictionaryItem[]} modlist 模块列表
    * @param {string} modType 模块名称
    * @param {*} originitem 初始化数据
    * @param {object} option 设置项
@@ -528,7 +528,7 @@ class DictionaryList extends DefaultDataWithLife {
   /**
    * 基于formdata和模块列表返回编辑完成的数据
    * @param {object} formData form数据
-   * @param {DictionaryData[]} modlist 模块列表
+   * @param {DictionaryItem[]} modlist 模块列表
    * @param {string} modType modType
    * @returns {object}
    */